Upgrade vSphere Hosts
Automated the upgrade of vSphere hosts with VMware Update Manager. This script asks for a cluster and image baseline for install. The script then grabs all hosts in the selected cluster and upgrades the hosts in the cluster one at a time. This is useful for environments where change control may not allow multiple hosts to be in maintenance mode at the same time.

The cluster variable accepts pipeline data and places multiple cluster names into an object so multiple clusters may be upgraded in a single run if desired.

Configuration backups are taken from each host and placed in the desired location.

The script grabs the desired cluster in vCenter and attempts to disconnect all ISO media to ensure they cannot prevent maintenance mode. Once complete, the desired baseline is applied and tested. Only hosts with a successful backup and non-compliant will be upgraded.

There is no error checking so use this script at your own risk.  All of the PowerCLI commands fail appropriately without error checking.